# vlo-accounts-frontend/src/vlo-accounts-client/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "ASP.NETCore_suvlo",
    "contact": {
      "name": "GJusz",
      "email": "gkjuszczyk@gmail.com"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/Auth/ExternalLogin/Callback": {
      "get": {
        "tags": [
          "ExternalLogin"
        ],
        "summary": "Callback executed when external auth provider redirects back to app",
        "parameters": [
          {
            "name": "returnUrl",
            "in": "query",
            "description": "",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "remoteError",
            "in": "query",
            "description": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Auth/ExternalLogin/CreateAccount": {
      "post": {
        "tags": [
          "ExternalLogin"
        ],
        "summary": "Endpoint used for creating accounts when there's no account associated with external login",
        "parameters": [
          {
            "name": "returnUrl",
            "in": "query",
            "description": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VLO_BOARDS.Areas.Auth.ExternalLoginRegisterInputModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/VLO_BOARDS.Areas.Auth.ExternalLoginRegisterInputModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/VLO_BOARDS.Areas.Auth.ExternalLoginRegisterInputModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Auth/Login": {
      "post": {
        "tags": [
          "Login"
        ],
        "summary": "Endpoint used to log user in based on username, password",
        "parameters": [
          {
            "name": "returnUrl",
            "in": "query",
            "description": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VLO_BOARDS.Areas.Auth.LoginInputModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/VLO_BOARDS.Areas.Auth.LoginInputModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/VLO_BOARDS.Areas.Auth.LoginInputModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/VLO_BOARDS.Areas.Auth.LoginResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VLO_BOARDS.Areas.Auth.LoginResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/VLO_BOARDS.Areas.Auth.LoginResult"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "VLO_BOARDS.Areas.Auth.ExternalLoginRegisterInputModel": {
        "required": [
          "email",
          "username"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "username": {
            "type": "string",
            "format": "string"
          }
        },
        "additionalProperties": false
      },
...
      "VLO_BOARDS.Areas.Auth.LoginInputModel": {
        "required": [
          "captchaResponse",
          "password",
          "rememberMe",
          "username"
        ],
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "format": "string"
          },
          "password": {
            "type": "string",
            "format": "password"
          },
          "rememberMe": {
            "type": "boolean"
          },
          "captchaResponse": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "VLO_BOARDS.Areas.Auth.LoginResult": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "nullable": true
          },
          "redirect": {
            "type": "boolean"
          },
          "returnUrl": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}